refactor(solver): keep #9709 fix within under-cap shards
Reshape the indexed-access error-type fix so it does not modify any
over-2000-line shard (AGENTS §19), addressing the review blocker:

- Move the conditional error-type propagation out of the over-cap
  `conditional.rs` into the under-cap `evaluate/support.rs` dispatch
  (`visit_conditional`), which is the primary conditional evaluation entry.
- Relocate the two affected solver micro-tests out of the over-cap
  `evaluate_tests.rs` / `operations_tests.rs` into a focused under-cap
  `indexed_access_error_type_tests.rs` shard, with tsc-correct error-type
  assertions.

Behavior is unchanged; the core fix stays in the under-cap `index_access.rs`.
